gap_test.py: Cleanup and refactors
Testing with --one-side-skip was 20% slower than expected.
Lots of debugging later it's not clear. call to prev_prime at P=1511
was slower than call to next_prime. Ultimately --taskset seemed to
"fix" the problem but it was a frustrating 2 days.

Maybe some stats should be saved to suggest --taskset?
Maybe should warn on P <= 2000?
One take away is that maybe commit should be called on a random fraction
on time?