combined_sieve: Repeat large_primes optimizations for medium_primes
These are similiar improvements to avoiding wide memory read/writes.
When prime p divides number n, then n is composite BUT n might also
be already known as a composite from a trivial prime (2, 3, 5, 7) in
that case we can skip marking n as composite (a second time) and
avoid a wide memory write.